requirements.txt
fastapi 
uvicorn 
pytsqlalchemy